+++ src/test/java/com/epam/reportportal/testng/ParametersBypassTest.java
@@ -83,7 +83,7 @@ public class ParametersBypassTest {
 
 		methodStarts.stream().map(StartTestItemRQ::getParameters).peek(pl -> assertThat(pl, hasSize(2))).forEach(testParams -> {
 			assertThat(testParams, hasSize(2));
-			assertThat(testParams.get(0).getKey(), equalTo(Integer.class.getCanonicalName()));
+			assertThat(testParams.get(0).getKey(), equalTo(Integer.TYPE.getCanonicalName()));
 			assertThat(testParams.get(1).getKey(), equalTo(String.class.getCanonicalName()));
 
 			assertThat(testParams.get(0).getValue() + "-" + testParams.get(1).getValue(), anyOf(equalTo("1-one"), equalTo("2-two")));
@@ -92,7 +92,8 @@ public class ParametersBypassTest {
 
 	private static void verify_null_parameter(List<StartTestItemRQ> methodStarts) {
 		assertThat(methodStarts, hasSize(3));
-		List<String> expectedKeys = Arrays.asList(String.class.getCanonicalName(), String.class.getCanonicalName(), "arg0");
+		String strClassName = String.class.getCanonicalName();
+		List<String> expectedKeys = Collections.nCopies(3, strClassName);
 		List<String> expectedValues = Arrays.asList("one", "two", "NULL");
 		List<String> expectedResult = IntStream.range(0, 3)
 				.mapToObj(i -> expectedKeys.get(i) + "-" + expectedValues.get(i))
@@ -127,7 +128,7 @@ public class ParametersBypassTest {
 	}
 
 	@Test
-	public void verify_parameter_key_annotation_bypass() {
+	public void verify_parameter_key_annotation_bypass_for_test_method() {
 		TestUtils.runTests(Collections.singletonList(TestReportPortalListener.class), ParameterNamesTest.class);
 
 		ArgumentCaptor<StartTestItemRQ> testCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);