+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [5.0.4-TESTNG6]
 ### Changed
 - `@NotNull` annotation replaced with `@Nonnull`
 - Test parameter processing logic was moved into the client and updated. Now it supports `@parameterKey` annotations for factory